Complete CC: Pen Testing Room
Completed the CC: Pen Testing room as recommended by the free path.

Section 1 included introductions to Network Utilities such as nmap
and Netcat.

Section 2 introduced Web Enumeration and tools like gobuster and
nikto.

Section 3 was about Metasploit.

Section 4 taught Hash Cracking and using hashcat and John The Ripper.

Section 5 included SQL Injection and an introduction to sqlmap.

Section 6 was an introduction to Samba, though, no use of the tools
or any machines to play with.

Section 7 was a final exam of the rooms. Though I completed this,
and am pushing up just this checkmark, a write-up for the final room
will come tonight.