Sublist3r and Metasploit
The sublist3r room took me through the installation, optional flags, and
a look at the output from a scan. The scan was pre-run, and a downloaded
file so I didn't have to wait for a scan.

Metasploit was a more in depth look at this tool. From initializing the
database, to understanding the available modules and commands within
msfconsole. Then there was a machine to exploit, and I went through some
steps that would take place after acquiring access.